+++ examples/use_cookie/src/main.rs
@@ -1,7 +1,7 @@
 use leptos::*;
 use leptos_use::docs::demo_or_body;
 use leptos_use::use_cookie;
-use leptos_use::utils::FromToStringCodec;
+use codee::string::FromToStringCodec;
 use rand::prelude::*;
 
 #[component]