+++ src/use_cookie.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::too_many_arguments)]
 
 use crate::core::now;
+use crate::utils::get_header;
 use codee::{CodecError, Decoder, Encoder};
 use cookie::time::{Duration, OffsetDateTime};
 pub use cookie::SameSite;
@@ -11,7 +12,7 @@ use std::rc::Rc;
 
 /// SSR-friendly and reactive cookie access.
 ///
-/// You can use this function multiple times in your for the same cookie and they're signals will synchronize
+/// You can use this function multiple times for the same cookie and their signals will synchronize
 /// (even across windows/tabs). But there is no way to listen to changes to `document.cookie` directly so in case
 /// something outside of this function changes the cookie, the signal will **not** be updated.
 ///
@@ -57,10 +58,10 @@ use std::rc::Rc;
 /// ```
 ///
 /// Values are (en)decoded via the given codec. You can use any of the string codecs or a
-/// binary codec wrapped in [`Base64`].
+/// binary codec wrapped in `Base64`.
 ///
 /// > Please check [the codec chapter](https://leptos-use.rs/codecs.html) to see what codecs are
-///   available and what feature flags they require.
+/// > available and what feature flags they require.
 ///
 /// ## Cookie attributes
 ///
@@ -92,14 +93,14 @@ use std::rc::Rc;
 /// the HTTP response header according to options (if provided).
 /// The returned `WriteSignal` may not affect the cookie headers on the server! It will try and write
 /// the headers buy if this happens after the headers have already been streamed to the client then
-/// this will have no effect. 
+/// this will have no effect.
 ///
 /// > If you're using `axum` you have to enable the `"axum"` feature in your Cargo.toml.
 /// > In case it's `actix-web` enable the feature `"actix"`, for `spin` enable `"spin"`.
 ///
 /// ### Bring your own header
 ///
-/// In case you're neither using Axum nor Actix, or the default implementation is not to your liking,
+/// In case you're neither using Axum, Actix nor Spin, or the default implementation is not to your liking,
 /// you can provide your own way of reading and writing the cookie header value.
 ///
 /// ```
@@ -136,10 +137,6 @@ use std::rc::Rc;
 /// # view! {}
 /// # }
 /// ```
-///
-/// ## Create Your Own Custom Codec
-///
-/// All you need to do is to implement the [`StringCodec`] trait together with `Default` and `Clone`.
 pub fn use_cookie<T, C>(cookie_name: &str) -> (Signal<Option<T>>, WriteSignal<Option<T>>)
 where
     C: Encoder<T, Encoded = String> + Decoder<T, Encoded = str>,
@@ -363,6 +360,8 @@ where
     #[cfg(feature = "ssr")]
     {
         if !readonly {
+            let cookie_name = cookie_name.to_owned();
+
             create_isomorphic_effect(move |_| {
                 let value = cookie
                     .with(|cookie| {
@@ -373,20 +372,27 @@ where
                         })
                     })
                     .flatten();
-                jar.update_value(|jar| {
-                    write_server_cookie(
-                        cookie_name,
-                        value,
-                        jar,
-                        max_age,
-                        expires,
-                        domain,
-                        path,
-                        same_site,
-                        secure,
-                        http_only,
-                        ssr_set_cookie,
-                    )
+
+                jar.update_value({
+                    let domain = domain.clone();
+                    let path = path.clone();
+                    let ssr_set_cookie = Rc::clone(&ssr_set_cookie);
+
+                    |jar| {
+                        write_server_cookie(
+                            &cookie_name,
+                            value,
+                            jar,
+                            max_age,
+                            expires,
+                            domain,
+                            path,
+                            same_site,
+                            secure,
+                            http_only,
+                            ssr_set_cookie,
+                        )
+                    }
                 });
             });
         }
@@ -491,78 +497,7 @@ impl<T, E, D> Default for UseCookieOptions<T, E, D> {
             path: None,
             same_site: None,
             ssr_cookies_header_getter: Rc::new(move || {
-                #[cfg(feature = "ssr")]
-                {
-                    #[cfg(all(feature = "actix", feature = "axum"))]
-                    compile_error!("You can only enable one of features \"actix\" and \"axum\" at the same time");
-
-                    #[cfg(all(feature = "actix", feature = "spin"))]
-                    compile_error!("You can only enable one of features \"actix\" and \"spin\" at the same time");
-
-                    #[cfg(all(feature = "axum", feature = "spin"))]
-                    compile_error!("You can only enable one of features \"axum\" and \"spin\" at the same time");
-
-                    #[cfg(feature = "actix")]
-                    const COOKIE: http0_2::HeaderName = http0_2::header::COOKIE;
-                    #[cfg(any(feature = "axum", feature = "spin"))]
-                    const COOKIE: http1::HeaderName = http1::header::COOKIE;
-
-                    #[cfg(feature = "actix")]
-                    type HeaderValue = http0_2::HeaderValue;
-                    #[cfg(feature = "axum")]
-                    type HeaderValue = http1::HeaderValue;
-
-                    #[cfg(any(feature = "axum", feature = "actix", feature = "spin"))]
-                    let headers;
-                    #[cfg(feature = "actix")]
-                    {
-                        headers = use_context::<actix_web::HttpRequest>()
-                            .map(|req| req.headers().clone());
-                    }
-                    #[cfg(feature = "axum")]
-                    {
-                        headers = use_context::<http1::request::Parts>().map(|parts| parts.headers);
-                    }
-                    #[cfg(feature = "spin")]
-                    {
-                        headers = use_context::<leptos_spin::RequestParts>()
-                            .map(|parts| parts.headers().clone());
-                    }
-
-                    #[cfg(all(
-                        not(feature = "axum"),
-                        not(feature = "actix"),
-                        not(feature = "spin")
-                    ))]
-                    {
-                        leptos::logging::warn!("If you're using use_cookie without the feature `axum`, `actix` or `spin` enabled, you should provide the option `ssr_cookies_header_getter`");
-                        None
-                    }
-
-                    #[cfg(any(feature = "axum", feature = "actix"))]
-                    {
-                        headers.map(|headers| {
-                            headers
-                                .get(COOKIE)
-                                .cloned()
-                                .unwrap_or_else(|| HeaderValue::from_static(""))
-                                .to_str()
-                                .unwrap_or_default()
-                                .to_owned()
-                        })
-                    }
-                    #[cfg(feature = "spin")]
-                    {
-                        headers.and_then(|headers| {
-                            headers
-                                .iter()
-                                .find(|(key, _)| **key == COOKIE)
-                                .and_then(|(_, value)| String::from_utf8(value.to_vec()).ok())
-                        })
-                    }
-                }
-                #[cfg(not(feature = "ssr"))]
-                None
+                get_header!(COOKIE, use_cookie, ssr_cookies_header_getter)
             }),
             ssr_set_cookie: Rc::new(|cookie: &Cookie| {
                 #[cfg(feature = "ssr")]