+++ src/use_raf_fn.rs
@@ -89,7 +89,7 @@ pub fn use_raf_fn_with_options(
         let previous_frame_timestamp = Cell::new(0.0_f64);
 
         move |timestamp: f64| {
-            if !is_active.get_untracked() {
+            if !is_active.try_get_untracked().unwrap_or_default() {
                 return;
             }
 
@@ -165,6 +165,6 @@ pub struct UseRafFnCallbackArgs {
     /// Time elapsed between this and the last frame.
     pub delta: f64,
 
-    /// Time elapsed since the creation of the web page. See [MDN Docs](https://developer.mozilla.org/en-US/docs/Web/API/DOMHighResTimeStamp#the_time_origin Time origin).
+    /// Time elapsed since the creation of the web page. See [MDN Docs](https://developer.mozilla.org/en-US/docs/Web/API/DOMHighResTimeStamp#the_time_origin) Time origin.
     pub timestamp: f64,
 }